+++ src/test/kotlin/com/zepben/evolve/streaming/grpc/ConnectTest.kt
@@ -18,7 +18,6 @@ import org.hamcrest.Matchers.equalTo
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import java.net.http.HttpClient
 
 internal class ConnectTest {